+++ BenchmarkTests/DataUpload/DataUploaderBenchmarkTests.swift
@@ -27,7 +27,7 @@ class DataUploaderBenchmarkTests: BenchmarkTests {
     /// `DataUploader` leaves no memory footprint (the memory peak after upload is less or equal `0kB`).
     func testUploadingDataToServer_leavesNoMemoryFootprint() throws {
         let dataUploader = DataUploader(
-            httpClient: HTTPClient(),
+            httpClient: URLSessionClient(),
             requestBuilder: FeatureRequestBuilderMock()
         )