+++ DatadogRUM/Sources/RUMMonitor/Scopes/RUMUserActionScope.swift
@@ -141,7 +141,7 @@ internal class RUMUserActionScope: RUMScope, RUMContextProvider {
             dd: .init(
                 action: nil,
                 browserSdkVersion: nil,
-                configuration: nil,
+                configuration: .init(sessionReplaySampleRate: nil, sessionSampleRate: Double(dependencies.sessionSampler.samplingRate)),
                 session: .init(plan: .plan1)
             ),
             action: .init(