+++ IntegrationTests/IntegrationScenarios/Scenarios/CrashReporting/CrashReportingWithLoggingScenarioTests.swift
@@ -97,7 +97,7 @@ class CrashReportingWithLoggingScenarioTests: IntegrationTests, LoggingCommonAss
         )
 
         // Assert other characteristics important for crash reporting
-        crashLog.assertServiceName(equals: "ui-tests-service-name")
+        crashLog.assertService(equals: "ui-tests-service-name")
         crashLog.assertLoggerName(equals: "crash-reporter")
     }
 }