+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Unreleased
 
+- [BUGFIX] Manual trace injection APIs are not available in DatadogTrace. See [#1415][].
+- [BUGFIX] Fix session replay uploads to AP1 site. See [#1418][].
+- [BUGFIX] Allow instantiating custom instance of the SDK after default one. See [#1413][].
+- [BUGFIX] Do not propagate attributes from Errors and LongTasks to Views.
+- [IMPROVEMENT] Upgrade to PLCrashReporter 1.11.1.
+- [FEATURE] Report session sample rate to the backend with RUM events. See [#1410][]
+- [IMPROVEMENT] Expose Session Replay to Objective-C. see [#1419][]
+
 # 2.0.0 / 31-07-2023
 
 Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATION.md) to upgrade from `1.x` versions.
@@ -488,6 +496,11 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1331]: https://github.com/DataDog/dd-sdk-ios/pull/1331
 [#1328]: https://github.com/DataDog/dd-sdk-ios/pull/1328
 [#1355]: https://github.com/DataDog/dd-sdk-ios/pull/1355
+[#1410]: https://github.com/DataDog/dd-sdk-ios/pull/1410
+[#1413]: https://github.com/DataDog/dd-sdk-ios/pull/1413
+[#1415]: https://github.com/DataDog/dd-sdk-ios/pull/1415
+[#1418]: https://github.com/DataDog/dd-sdk-ios/pull/1418
+[#1419]: https://github.com/DataDog/dd-sdk-ios/pull/1419
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu